M4A changes

+++ java/commoncomponent/common.serialization.deserialization/src/main/java/net/atos/daf/ct2/pojo/standard/MonitorDocument.java
@@ -27,7 +27,7 @@ public class MonitorDocument implements Serializable {
     private ROEventObject roEventObject;
 
     @JsonProperty(value = "GPSHdop")
-    private Integer gpsHdop;
+    private Double gpsHdop;
     @JsonProperty(value = "GPSSpeed")
     private Long gpsSpeed;
